third_party/blink/web_tests/fast/dom/Window/window-open-no-multiple-windows-expected.txt

CONSOLE ERROR: line 17: Not allowed to navigate top frame to data URL: data:text/html;charset=utf-8,<html><body>The test passes if this page opens in the same window</body></html>
PASS: Popup opened in the same window
